@@ -1,6 +1,6 @@
 {
   "name": "react-form-controlled",
-  "version": "0.9.29",
+  "version": "0.9.30",
   "description": "React controlled form components. The main idea is to make forms as simple as possible.",
   "author": {
     "name": "Zlatko Fedor",
@@ -30,7 +30,7 @@
   "scripts": {
     "prepublish": "npm run build",
     "test": "babel-node ./node_modules/gulp/bin/gulp.js test",
-    "build": "node ./node_modules/gulp/bin/gulp.js build",
+    "build": "babel-node ./node_modules/gulp/bin/gulp.js build",
     "eslint": "node ./node_modules/eslint/bin/eslint.js --ext .js,.jsx ./src/"
   },
   "dependencies": {
@@ -42,14 +42,16 @@
   "devDependencies": {
     "babel": "^5.8.28",
     "babel-eslint": "^4.1.4",
-    "eslint": "^1.9.0",
+    "eslint": "^1.10.0",
     "eslint-config-airbnb": "^0.1.0",
     "eslint-loader": "^1.1.1",
-    "eslint-plugin-react": "^3.8.0",
+    "eslint-plugin-react": "^3.9.0",
     "gulp": "^3.9.0",
     "gulp-babel": "^5.2.1",
-    "gulp-mocha": "^2.1.3",
-    "should": "^7.1.1"
+    "gulp-mocha": "^2.2.0",
+    "should": "^7.1.1",
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3"
   },
   "peerDependencies": {
     "react": "0.14.x"
